Fix prop drilling
