+++ src/components/Feed/CreatePost.jsx
@@ -3,8 +3,10 @@ import UserAvatar from '../atoms/UserAvatar'
 import { loggedInUser } from '../../data/userData'
 import IconText from '../molecules/IconText'
 import Icon from '../atoms/Icon'
+import { connect } from 'react-redux'
+import { addNewPost } from '../../redux/actions'
 
-export default function CreatePost( {onNewPostCreated} ) {
+function CreatePost( {addNewPost} ) {
     const [isModalOpen, setIsModalOpen] = useState(false);
     const [postText, setPostText] = useState('');
     const [mediaFile, setMediaFile] = useState(null);
@@ -44,7 +46,7 @@ export default function CreatePost( {onNewPostCreated} ) {
             comments: []
         };
         
-        onNewPostCreated(newPost);
+        addNewPost(newPost);
         toggleModal();
     }
 
@@ -110,3 +112,9 @@ export default function CreatePost( {onNewPostCreated} ) {
         </>
     )
 }
+
+const mapDispatchToProps = {
+    addNewPost,
+};
+
+export default connect(null, mapDispatchToProps)(CreatePost);
