+++ src/components/Feed/Feed.jsx
@@ -4,16 +4,12 @@ import CreatePost from './CreatePost'
 import PostList from './PostList';
 import DateFilter from './DateFilter';
 
-export default function Feed({posts, onNewPostCreated, startDate, endDate, handleStartDateChange, handleEndDateChange, 
-    isLoading, loaderRef}) {
+export default function Feed({ loaderRef }) {
     return (
         <div className='flex-col feed'>
-            <CreatePost onNewPostCreated={onNewPostCreated}/>
-            <DateFilter startDate={startDate} 
-            endDate={endDate}
-            handleStartDateChange = {handleStartDateChange} 
-            handleEndDateChange = {handleEndDateChange}/>
-            <PostList posts={posts} isLoading = {isLoading} loaderRef = {loaderRef}/>
+            <CreatePost />
+            <DateFilter />
+            <PostList loaderRef = {loaderRef}/>
         </div>
     )
 }
