+++ src/components/Feed/PostList.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import Post from './Post'
 import BeatLoader from "react-spinners/BeatLoader";
+import { connect } from 'react-redux';
 
-export default function PostList( {posts, isLoading, loaderRef} ) {
+function PostList( {posts, isLoading, loaderRef} ) {
     return (
         <>
             <div className='post-list flex-col'>
@@ -31,3 +32,10 @@ export default function PostList( {posts, isLoading, loaderRef} ) {
         
     )
 }
+
+const mapStateToProps = (state) => ({
+    posts: state.posts.posts,
+    isLoading: state.posts.isLoading,
+});
+
+export default connect(mapStateToProps)(PostList);
