+++ src/components/Header/SearchBar.jsx
@@ -1,13 +1,15 @@
 import React, { useEffect } from "react";
 import Icon from "../atoms/Icon";
 import UserList from "./UserList";
+import { connect } from "react-redux";
+import { setSearchQuery, toggleUserList } from "../../redux/actions";
 
-export default function SearchBar( {searchQuery, showUserList, userList, onSearchQueryChange, onShowUserListChange, onSearchPostsByUserId} ) {
+function SearchBar( {searchQuery, showUserList, setSearchQuery, toggleUserList} ) {
     function handleBlur(){
         // clicking on IconText in UserList fires onBlur for Input element when Searching post by user name/id. 
         // On onBlur the UserList is lost and handleSearchByUserName is not propogated. 
         setTimeout(() => {
-            onShowUserListChange(false);
+            toggleUserList(false);
         }, 200);
     }
 
@@ -16,15 +18,26 @@ export default function SearchBar( {searchQuery, showUserList, userList, onSearc
             <div className="search-input">
                 <Icon iconName='search' size='16px'/>
                 <input type="text" placeholder="Search" value={searchQuery} 
-                onChange={(e) => onSearchQueryChange(e.target.value)} 
-                onFocus={() => onShowUserListChange(true)}
+                onChange={(e) => setSearchQuery(e.target.value)} 
+                onFocus={() => toggleUserList(true)}
                 onBlur={handleBlur}/> 
             </div>
 
             {showUserList && (
-                <UserList userList={userList}
-                onSearchPostsByUserId={onSearchPostsByUserId}/>
+                <UserList />
             )}
         </div>
     );
 }
+
+const mapStateToProps = (state) => ({
+    searchQuery: state.search.searchQuery,
+    showUserList: state.user.showUserList,
+});
+
+const mapDispatchToProps = {
+    setSearchQuery,
+    toggleUserList,
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(SearchBar);
